pyproject.toml
[project]
name = "security-edu"
version = "0.1.0"
description = "Аналитика и аналитическая инфраструктура в сфере информационной безопасности"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "duckdb>=1.4.3",
    "ipykernel>=7.1.0",
    "polars>=1.37.0",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "scikit-learn>=1.8.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]
